Make TextSpan u32 to match the compiler one (#1748)
commit-id:107b5e17

---

**Stack**:
- #1749
- #1748 ⬅

⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*

+++ plugins/cairo-lang-macro-stable/src/lib.rs
@@ -16,8 +16,8 @@ pub struct StableToken {
 #[repr(C)]
 #[derive(Debug)]
 pub struct StableTextSpan {
-    pub start: usize,
-    pub end: usize,
+    pub start: u32,
+    pub end: u32,
 }
 
 #[repr(C)]
